backend/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.6.0
djangorestframework==3.15.2
Pillow==11.1.0
dj-database-url==2.1.0
gunicorn==23.0.0
whitenoise==6.8.2
